Add initial license and basic description in a README.rst
This license is the same one that saltstack-formulas uses and should be seen
as a good starting point. If you have good reason for a different license,
let us know.

The README is just the bare minimum. We will flesh it out more as the repo's grow
and we see what kind of contributions we have.